+++ sympy/codegen/ast.py
@@ -190,6 +190,99 @@ def aug_assign(lhs, op, rhs):
         raise ValueError("Unrecognized operator %s" % op)
     return Relational.ValidRelationOperator[op + '='](lhs, rhs)
 
+
+class CodeBlock(Basic):
+    def __new__(cls, *assignments):
+        if not all(isinstance(i, Assignment) for i in assignments):
+            # Will support more things later
+            raise TypeError("CodeBlock inputs must be Assignments")
+
+        left_hand_sides = []
+        right_hand_sides = []
+        for i in assignments:
+            lhs, rhs = i.args
+            if lhs in left_hand_sides:
+                raise NotImplementedError("Duplicate assignments to the same "
+                "variable are not yet supported (%s)" % lhs)
+            left_hand_sides.append(lhs)
+            right_hand_sides.append(rhs)
+
+        obj = Basic.__new__(cls, *assignments)
+
+        obj.left_hand_sides = Tuple(*left_hand_sides)
+        obj.right_hand_sides = Tuple(*right_hand_sides)
+
+        return obj
+
+    @classmethod
+    def topological_sort(cls, assignments):
+        """
+        Return a CodeBlock with topologically sorted assignments so that
+        variables are assigned before they are used.
+
+        The existing order of assignments is preserved as much as possible.
+
+        This is a class constructor so that the default constructor for
+        CodeBlock can error when variables are used before they are assigned.
+        """
+        from sympy.utilities.iterables import topological_sort
+        # Create a graph where the nodes are assignments and there is a directed edge
+        # between nodes that use a variable and nodes that assign that
+        # variable, like
+
+        # [(x := 1, y := x + 1), (x := 1, z := y + z), (y := x + 1, z := y + z)]
+
+        # If we then topologically sort these nodes, they will be in
+        # assignment order, like
+
+        # x := 1
+        # y := x + 1
+        # z := y + z
+
+        # The nodes
+        #
+        # enumerate keeps nodes in the same order they are already in if
+        # possible. It will also allow us to handle duplicate assignments to
+        # the same variable when those are implemented.
+        A = list(enumerate(assignments))
+
+        # {variable: [assignments using variable]}
+        # like {x: [y := x + 1, z := y + x], ...}
+        var_map = {}
+
+        # Edges in the graph
+        E = []
+        for i in A:
+            if i[1].lhs in var_map:
+                E.append((var_map[i[1].lhs], i))
+            var_map[i[1].lhs] = i
+        for i in A:
+            for x in i[1].rhs.free_symbols:
+                if x not in var_map:
+                    raise ValueError("Undefined variable %s" % x)
+                E.append((var_map[x], i))
+
+        ordered_assignments = topological_sort([A, E])
+        # De-enumerate the result
+        return cls(*list(zip(*ordered_assignments))[1])
+
+    def cse(self, symbols=None, optimizations=None, postprocess=None,
+        order='canonical'):
+        """
+        Return a new code block with common subexpressions eliminated
+
+        See the docstring of :func:`sympy.simplify.cse_main.cse` for more information.
+        """
+        # TODO: Check that the symbols are new
+        from sympy.simplify.cse_main import cse
+        replacements, reduced_exprs = cse(self.right_hand_sides, symbols=symbols,
+            optimizations=optimizations, postprocess=postprocess, order=order)
+        assert len(reduced_exprs) == 1
+        new_block = tuple(Assignment(var, expr) for var, expr in
+            zip(self.left_hand_sides, reduced_exprs[0]))
+        new_assignments = tuple(Assignment(*i) for i in replacements)
+        return self.topological_sort(new_assignments + new_block)
+
 class For(Basic):
     """Represents a 'for-loop' in the code.
